+++ CHANGELOG.md
@@ -11,6 +11,13 @@ adheres to [Semantic Versioning][semver].
 
 ## [Unreleased]
 
+### Added
+
+* Added initial WebSocket support. `gocurl` now supports `ws://` and `wss://`
+  URLs. `-d` can be used to specify initial data to send.  ([#17][#17])
+
+[#15]: https://github.com/ameshkov/gocurl/issues/15
+
 [unreleased]: https://github.com/ameshkov/gocurl/compare/v1.3.0...HEAD
 
 ## [1.3.0] - 2023-09-23